REFACTOR downloading of data
