+++ WalletLess.xcodeproj/xcuserdata/fkjamilov.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -3,4 +3,102 @@
    uuid = "3AC2C061-A7A3-449C-ADA5-23A9FB1A4016"
    type = "1"
    version = "2.0">
+   <Breakpoints>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            uuid = "A1B2719B-2591-4137-9D79-846731E2F000"
+            shouldBeEnabled = "Yes"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "WalletLess/Utils/RealmManager.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "59"
+            endingLineNumber = "59"
+            landmarkName = "addMerhant(name:image:locations:scannedCode:typeOfCode:)"
+            landmarkType = "7">
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            uuid = "D7850812-59DE-4045-B021-770858DEB165"
+            shouldBeEnabled = "Yes"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "WalletLess/Utils/RealmManager.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "36"
+            endingLineNumber = "36"
+            landmarkName = "addMerhant(name:image:locations:scannedCode:typeOfCode:)"
+            landmarkType = "7">
+            <Locations>
+               <Location
+                  uuid = "D7850812-59DE-4045-B021-770858DEB165 - 7796254ad40e9829"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "WalletLess.RealmManager.addMerhant(name: Swift.String, image: Swift.String, locations: RealmSwift.List&lt;WalletLess.LocationsDto&gt;, scannedCode: Swift.String, typeOfCode: WalletLess.CodeType) -&gt; ()"
+                  moduleName = "WalletLess"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/fkjamilov/Workspace/WalletLess/WalletLess/Utils/RealmManager.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "34"
+                  endingLineNumber = "34"
+                  offsetFromSymbolStart = "173">
+               </Location>
+               <Location
+                  uuid = "D7850812-59DE-4045-B021-770858DEB165 - 6e2b1bd06fa7c919"
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "closure #1 (Swift.Optional&lt;Foundation.URL&gt;, Swift.Optional&lt;Swift.Error&gt;) -&gt; () in WalletLess.RealmManager.addMerhant(name: Swift.String, image: Swift.String, locations: RealmSwift.List&lt;WalletLess.LocationsDto&gt;, scannedCode: Swift.String, typeOfCode: WalletLess.CodeType) -&gt; ()"
+                  moduleName = "WalletLess"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/fkjamilov/Workspace/WalletLess/WalletLess/Utils/RealmManager.swift"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "35"
+                  endingLineNumber = "35"
+                  offsetFromSymbolStart = "483">
+               </Location>
+            </Locations>
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            uuid = "ED0AB027-91C2-44E0-AAB0-E51A057FF57E"
+            shouldBeEnabled = "Yes"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "WalletLess/Utils/RealmManager.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "31"
+            endingLineNumber = "31"
+            landmarkName = "addMerhant(name:image:locations:scannedCode:typeOfCode:)"
+            landmarkType = "7">
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            uuid = "AC29B808-8474-4FB9-8990-875BA76EE81E"
+            shouldBeEnabled = "Yes"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "WalletLess/Utils/RealmManager.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "35"
+            endingLineNumber = "35"
+            landmarkName = "addMerhant(name:image:locations:scannedCode:typeOfCode:)"
+            landmarkType = "7">
+         </BreakpointContent>
+      </BreakpointProxy>
+   </Breakpoints>
 </Bucket>
