+++ WalletLess/Gradiens/BackgroundGradient.swift
@@ -1,26 +1,29 @@
-//  Created by Filip Kjamilov on 24.5.22.
+//  Created by Filip Kjamilov on 18.10.22.
 
 import SwiftUI
 
-public struct CardGradientBackground: View {
+public struct GradientBackground: View {
     public var body: some View {
         ZStack {
             LinearGradient(colors: [Color.cyan.opacity(0.7), Color.purple.opacity(0.3)],
                            startPoint: .topLeading,
                            endPoint: .bottomTrailing)
             Circle()
-                .frame(width: 120)
-                .foregroundStyle(LinearGradient(colors: [Color.mint, Color.purple],
+                .frame(width: 300)
+                .foregroundStyle(LinearGradient(colors: [Color.mint.opacity(0.5), Color.purple.opacity(0.6)],
                                                 startPoint: .top,
                                                 endPoint: .leading))
-                .offset(x: -150, y: -60)
-            RoundedRectangle(cornerRadius: 20, style: .continuous)
-                .frame(width: 200, height: 200)
+//                .blur(radius: 10)
+                .offset(x: -100, y: -150)
+            RoundedRectangle(cornerRadius: 30, style: .continuous)
+                .frame(width: 500, height: 500)
                 .foregroundStyle(LinearGradient(colors: [Color.purple.opacity(0.6), Color.mint.opacity(0.5)],
                                                 startPoint: .top,
                                                 endPoint: .leading))
-                .offset(x: 180, y: -120)
-                .rotationEffect(.degrees(35))
+                .offset(x: 300)
+//                .blur(radius: 30)
+                .rotationEffect(.degrees(30))
+            
         }.ignoresSafeArea(.all)
     }
 }
