+++ WalletLess/Utils/LocationManager.swift
@@ -1,4 +1,4 @@
-//  Created by fkjamilov on 15.10.22.
+//  Created by Filip Kjamilov on 15.10.22.
 
 import Foundation
 import CoreLocation
