+++ WalletLess/Views/dashboard/Dashboard.swift
@@ -11,8 +11,9 @@ struct Dashboard: View {
     private var language = LocalizationService.shared.language
     
     @StateObject private var viewModel = DashboardViewModel()
-    
     @Binding var tabSelection: Int
+    
+    // TODO: FKJ - Check if we can remove this!
     @State private var currentMerchant: MerchantDto = MerchantDto()
     
     // MARK: - Location properties
@@ -35,30 +36,34 @@ struct Dashboard: View {
                 } else {
                     ScrollView {
                         ForEach(viewModel.realmManager.merchants.filter({ !$0.isInvalidated }), id: \.id) { merchant in
-                            
-                            Image(uiImage: UIImage(data: merchant.downloadedImage!)!)
-                                .resizable()
-                                .aspectRatio(contentMode: .fit)
-                                .cornerRadius(15)
-                                .padding(.all, 3)
-                                .padding(.leading, 7)
-                                .padding(.trailing, 7)
-                                .listRowInsets(.init())
-                                .onTapGesture {
-                                    currentMerchant = merchant
-                                    viewModel.isPresentingSheet.toggle()
-                                }
-                                .onLongPressGesture {
-                                    currentMerchant = merchant
-                                    viewModel.isPresentingConfirmationDialog = true
-                                }
-                                .alert("Are you sure you want to remove \(currentMerchant.name) from the list?", isPresented: $viewModel.isPresentingConfirmationDialog) {
-                                    Button("Confirm".localized(language), role: .destructive) {
-                                        viewModel.realmManager.deleteMerchant(id: currentMerchant.id)
-                                        currentMerchant = MerchantDto()
+                            if merchant.downloadedImage != Data() {
+                                Image(uiImage: UIImage(data: merchant.downloadedImage!)!)
+                                    .resizable()
+                                    .aspectRatio(contentMode: .fit)
+                                    .cornerRadius(15)
+                                    .padding(.all, 3)
+                                    .padding(.leading, 7)
+                                    .padding(.trailing, 7)
+                                    .listRowInsets(.init())
+                                    .onTapGesture {
+                                        currentMerchant = merchant
+                                        viewModel.isPresentingSheet.toggle()
                                     }
-                                    Button("Cancel".localized(language), role: .cancel) { /* no-op */ }
-                                }
+                                    .onLongPressGesture {
+                                        currentMerchant = merchant
+                                        viewModel.isPresentingConfirmationDialog = true
+                                    }
+                                    .alert("Are you sure you want to remove \(currentMerchant.name) from the list?", isPresented: $viewModel.isPresentingConfirmationDialog) {
+                                        Button("Confirm".localized(language), role: .destructive) {
+                                            viewModel.realmManager.deleteMerchant(id: currentMerchant.id)
+                                            currentMerchant = MerchantDto()
+                                        }
+                                        Button("Cancel".localized(language), role: .cancel) { /* no-op */ }
+                                    }
+                            } else {
+                                ProgressView()
+                            }
+                            
                         }
                     }
                 }
@@ -73,50 +78,12 @@ struct Dashboard: View {
         }
         .onAppear {
             /// Start observing location when user is on `Dashboard`.
-            viewModel.startSortingCards()
+            viewModel.viewAppeared()
         }
         .onDisappear {
             /// Stop observing location when user not on `Dashboard`.
-            viewModel.stopUpdatingLocation()
+            viewModel.viewDissapeared()
         }
-    }
-}
-
-func generateCode(from string: String, codeType: String?) -> UIImage {
-    let data = Data(string.utf8)
-    let context = CIContext()
-    let fallbackCodeType = "CICode128BarcodeGenerator"
-    
-    if let filter = CIFilter(name: codeType ?? fallbackCodeType) {
-        filter.setValue(data, forKey: "inputMessage")
-        
-        if let barcodeImage = filter.outputImage {
-            if let barcodeCGImage = context.createCGImage(barcodeImage, from: barcodeImage.extent) {
-                return UIImage(cgImage: barcodeCGImage)
-            }
-        }
-    }
-    
-    return UIImage(systemName: "xmark") ?? UIImage()
-}
-
-public struct GradientBackground: View {
-    public var body: some View {
-        ZStack {
-            LinearGradient(colors: [Color.cyan.opacity(0.7), Color.purple.opacity(0.3)], startPoint: .topLeading, endPoint: .bottomTrailing)
-            Circle()
-                .frame(width: 300)
-                .foregroundStyle(LinearGradient(colors: [Color.mint.opacity(0.5), Color.purple.opacity(0.6)], startPoint: .top, endPoint: .leading))
-                .blur(radius: 10)
-                .offset(x: -100, y: -150)
-            RoundedRectangle(cornerRadius: 30, style: .continuous)
-                .frame(width: 500, height: 500)
-                .foregroundStyle(LinearGradient(colors: [Color.purple.opacity(0.6), Color.mint.opacity(0.5)], startPoint: .top, endPoint: .leading))
-                .offset(x: 300)
-                .blur(radius: 30)
-                .rotationEffect(.degrees(30))
-            
-        }.ignoresSafeArea(.all)
     }
 }
